refactor(auth): 인증 transport session error 모듈 분리
